Include Cloud's config in install-config
We need to pass the cloud's config data to the MCO in order to be able
to configure the cloud provider openstack. This commit passes such data
as part of the platform configuration since MCO consumes install-config
to get the cluster conifguration.

As part of this commit, we're now also validating the user's input to
verify whether the cloud name passed is correct.