WIP Some progress on getting Ignition via IP
The LbFloatingIP is now used as the predictible address for API. It
points to the bootstrap node first, then is moved to the first master
upon bootstrap removal. As a consequence, the LbFloatingIP becomes
a mandatory parameter for the installer on OpenStack platform.

This patch also cleans up the network architecture a little bit by
removing the subnet complexity for the deployed nodes. There is now
only one subnet for all the provisioned nodes.