bootstrap: add switch-api-endpoint service
There is a potential cycle where the temporary bootstrap control plane
gets torn down and the API endpoint on the bootstrap node points to
itself (via the floating IP) rather than the masters.

The installer is waiting for the bootstrapping to be completed, but the
`progress` service is unable to send the bootstrap-complete event,
because the API server is no longer running on the bootstrap node (which
still has the FIP attached).

Therefore, the installer never runs the bootstrap destroy terraform
actions and the FIP is stuck on the bootstrap node.

This adds a new service that waits until the `bootkube` and `openshift`
services finish (just like `progress.service` does), but then creates an
`/etc/hosts` entry for the API endpoints so that `progress` can
communicate with the master control plane.